# api/openapi.yaml
...
components:
  responses:
    Error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
      description: Hard failure envelope (friendly message from the error catalog)
  schemas:
...
    AuditVerifyResult:
      properties:
        brokenAt:
          type: integer
        entries:
          type: integer
        head:
          type: string
        intact:
          type: boolean
        reason:
          type: string
      required:
      - entries
      - intact
      type: object
...
    CertInfoBatchItem:
      properties:
        error:
          $ref: '#/components/schemas/BatchItemError'
        index:
          description: position in the request
          type: integer
        output:
          $ref: '#/components/schemas/CertInfoResponse'
        status:
          enum:
          - ok
          - error
          - skipped
          type: string
      type: object
    CertInfoBatchRequest:
      description: A batch of CertInfoRequest items with batch-wide controls.
      properties:
        concurrency:
          description: max items in parallel (0 = driver pool size)
          type: integer
        items:
          items:
            $ref: '#/components/schemas/CertInfoRequest'
          type: array
        policy:
          description: error policy (default continue-on-error)
          enum:
          - continue-on-error
          - fail-fast
          type: string
      required:
      - items
      type: object
    CertInfoBatchResponse:
      properties:
        failed:
          type: integer
        results:
          items:
            $ref: '#/components/schemas/CertInfoBatchItem'
          type: array
        succeeded:
          type: integer
        total:
          type: integer
      type: object
    CertInfoRequest:
      properties:
        buildChain:
          type: boolean
        cert:
          contentEncoding: base64
          type: string
        claims:
          type: boolean
        encoding:
          enum:
          - pem
          - der
          - base64
          type: string
        key:
          $ref: '#/components/schemas/KeySpec'
        method:
          enum:
          - ocsp
          - crl
          type: string
        trustedCerts:
          items:
            $ref: '#/components/schemas/TrustedCert'
          type: array
        validate:
          type: boolean
      type: object
...
    ErrorEnvelope:
      properties:
        error:
          properties:
            action:
              type: string
            code:
              type: string
            kind:
              type: string
            message:
              type: string
          required:
          - kind
          - message
          type: object
      required:
      - error
      type: object
    ExtractBatchItem:
      properties:
        error:
          $ref: '#/components/schemas/BatchItemError'
        index:
          description: position in the request
          type: integer
        output:
          $ref: '#/components/schemas/ExtractResponse'
        status:
          enum:
          - ok
          - error
          - skipped
          type: string
      type: object
    ExtractBatchRequest:
      description: A batch of ExtractRequest items with batch-wide controls.
      properties:
        concurrency:
          description: max items in parallel (0 = driver pool size)
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ExtractRequest'
          type: array
        policy:
          description: error policy (default continue-on-error)
          enum:
          - continue-on-error
          - fail-fast
          type: string
      required:
      - items
      type: object
    ExtractBatchResponse:
      properties:
        failed:
          type: integer
        results:
          items:
            $ref: '#/components/schemas/ExtractBatchItem'
          type: array
        succeeded:
          type: integer
        total:
          type: integer
      type: object
    ExtractRequest:
      properties:
        data:
          contentEncoding: base64
          type: string
        format:
          enum:
          - cms
          - xml
          - wsse
          type: string
        signature:
          contentEncoding: base64
          type: string
      required:
      - format
      - signature
      type: object
...
    KeySpec:
      properties:
        inline:
          $ref: '#/components/schemas/InlineKey'
        keyId:
          type: string
        path:
          $ref: '#/components/schemas/PathKey'
        token:
          $ref: '#/components/schemas/TokenKey'
      type: object
...
    MultisignRequirement:
      properties:
        bin:
          type: string
        iin:
          type: string
        label:
          type: string
        role:
          type: string
      type: object
    MultisignSession:
      properties:
        collected:
          items:
            $ref: '#/components/schemas/MultisignSignature'
          type: array
        container:
          contentEncoding: base64
          type: string
        createdAt:
          format: date-time
          type: string
        detached:
          type: boolean
        document:
          contentEncoding: base64
          type: string
        expiresAt:
          format: date-time
          type: string
        format:
          enum:
          - cms
          - xml
          type: string
        id:
          type: string
        minSigners:
          type: integer
        required:
          items:
            $ref: '#/components/schemas/MultisignRequirement'
          type: array
        status:
          enum:
          - pending
          - complete
          - expired
          type: string
        subject:
          type: string
      required:
      - id
      - status
      - format
      - detached
      - collected
      - createdAt
      - expiresAt
      type: object
    MultisignSignature:
      properties:
        bin:
          type: string
        iin:
          type: string
        name:
          type: string
        roles:
          items:
            type: string
          type: array
        satisfies:
          type: integer
        signedAt:
          format: date-time
          type: string
      required:
      - signedAt
      - satisfies
      type: object
    MultisignSubmitRequest:
      description: The container as it stands after a participant co-signed it elsewhere.
        The service verifies it and counts the new signer; it never signs here.
      properties:
        container:
          description: base64 co-signed container
          format: byte
          type: string
      required:
      - container
      type: object
...
    RegistryResponse:
      properties:
        checkedAt:
          format: date-time
          type: string
        failed:
          type: integer
        indeterminate:
          type: integer
        invalid:
          type: integer
        rows:
          items:
            $ref: '#/components/schemas/RegistryRow'
          type: array
        total:
          type: integer
        valid:
          type: integer
      required:
      - checkedAt
      - total
      - valid
      - invalid
      - indeterminate
      - failed
      - rows
      type: object
...
    SignBatchItem:
      properties:
        error:
          $ref: '#/components/schemas/BatchItemError'
        index:
          description: position in the request
          type: integer
        output:
          $ref: '#/components/schemas/SignResponse'
        status:
          enum:
          - ok
          - error
          - skipped
          type: string
      type: object
    SignBatchRequest:
      description: A batch of SignRequest items with batch-wide controls.
      properties:
        concurrency:
          description: max items in parallel (0 = driver pool size)
          type: integer
        items:
          items:
            $ref: '#/components/schemas/SignRequest'
          type: array
        policy:
          description: error policy (default continue-on-error)
          enum:
          - continue-on-error
          - fail-fast
          type: string
      required:
      - items
      type: object
    SignBatchResponse:
      properties:
        failed:
          type: integer
        results:
          items:
            $ref: '#/components/schemas/SignBatchItem'
          type: array
        succeeded:
          type: integer
        total:
          type: integer
      type: object
    SignRequest:
      properties:
        data:
          contentEncoding: base64
          type: string
        dataPath:
          type: string
        dataUrl:
          type: string
        detached:
          type: boolean
        existingSignature:
          contentEncoding: base64
          type: string
        format:
          enum:
          - cms
          - xml
          - wsse
          type: string
        inputPem:
          type: boolean
        key:
          $ref: '#/components/schemas/KeySpec'
        noCheckCertTime:
          type: boolean
        nodeId:
          type: string
        outputPem:
          type: boolean
        parentNamespace:
          type: string
        parentNode:
          type: string
        policy:
          type: string
        tsaUrl:
          type: string
        withTimestamp:
          type: boolean
      required:
      - format
      - data
      - key
      type: object
...
    VerifyBatchItem:
      properties:
        error:
          $ref: '#/components/schemas/BatchItemError'
        index:
          description: position in the request
          type: integer
        output:
          $ref: '#/components/schemas/VerifyResponse'
        status:
          enum:
          - ok
          - error
          - skipped
          type: string
      type: object
    VerifyBatchRequest:
      description: A batch of VerifyRequest items with batch-wide controls.
      properties:
        concurrency:
          description: max items in parallel (0 = driver pool size)
          type: integer
        items:
          items:
            $ref: '#/components/schemas/VerifyRequest'
          type: array
        policy:
          description: error policy (default continue-on-error)
          enum:
          - continue-on-error
          - fail-fast
          type: string
      required:
      - items
      type: object
    VerifyBatchResponse:
      properties:
        failed:
          type: integer
        results:
          items:
            $ref: '#/components/schemas/VerifyBatchItem'
          type: array
        succeeded:
          type: integer
        total:
          type: integer
      type: object
    VerifyRegistryRequest:
      description: A set of documents to check, each optionally labeled with the caller's
        own reference.
      properties:
        concurrency:
          description: max items in parallel (0 = driver pool size)
          type: integer
        items:
          items:
            $ref: '#/components/schemas/RegistryItemRequest'
          type: array
        policy:
          description: error policy (default continue-on-error)
          enum:
          - continue-on-error
          - fail-fast
          type: string
      required:
      - items
      type: object
    VerifyRequest:
      properties:
        archive:
          type: boolean
        checkCertTime:
          type: boolean
        claims:
          type: boolean
        data:
          contentEncoding: base64
          type: string
        dataPath:
          type: string
        dataUrl:
          type: string
        detached:
          type: boolean
        explain:
          type: boolean
        extractContent:
          type: boolean
        format:
          enum:
          - cms
          - xml
          - wsse
          type: string
        inputPem:
          type: boolean
        receipt:
          type: boolean
        report:
          type: boolean
        responderUrl:
          type: string
        revocationCheck:
          type: boolean
        revocationMethod:
          type: string
        signature:
          contentEncoding: base64
          type: string
        trustedCerts:
          items:
            $ref: '#/components/schemas/TrustedCert'
          type: array
      required:
      - format
      - signature
      type: object
...
info:
  description: REST/JSON transport for the qoltanba digital-signature service (RK
    / Kalkan). One call handles one item. Binary fields are base64-encoded strings;
    times are RFC 3339. All request and response keys are lowerCamelCase. Generated
    from the Go types — do not edit by hand. The proprietary Kalkan library is bring-your-own
    (BYOL) and is not part of this service.
  title: qoltanba REST API
  version: v1
openapi: 3.1.0
paths:
...
  /audit/export:
    get:
      operationId: auditExport
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/x-ndjson:
              schema:
                type: string
          description: The journal, one signed entry per line
        "206":
          description: Partial content (range request)
        "503":
          $ref: '#/components/responses/Error'
      summary: Stream the journal as NDJSON for a SIEM (supports range requests and
        conditional fetches)
      tags:
      - audit
  /audit/verify:
    get:
      operationId: auditVerify
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResult'
          description: Chain intact
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResult'
          description: Chain broken — brokenAt names the first altered entry
        "503":
          $ref: '#/components/responses/Error'
      summary: Walk the journal's hash chain and report whether — and where — it was
        altered
      tags:
      - audit
...
  /cert/info/batch:
    post:
      operationId: certInfoBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertInfoBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertInfoBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Parse many certificates
      tags:
      - batch
...
  /extract/batch:
    post:
      operationId: extractBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Recover content from many attached signatures
      tags:
      - batch
...
  /multisign/sessions/{id}/signatures:
    post:
      operationId: multisignSubmit
      parameters:
      - description: session id
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultisignSubmitRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultisignSession'
          description: Session after counting the new signer
        "400":
          $ref: '#/components/responses/Error'
        "404":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
      summary: Submit the container after a participant co-signed it
      tags:
      - multisign
...
  /sign/batch:
    post:
      operationId: signBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: 'Sign many items (aggregated JSON, or NDJSON stream with Accept: application/x-ndjson)'
      tags:
      - batch
...
  /verify/batch:
    post:
      operationId: verifyBatch
      parameters:
      - description: 'stream results as NDJSON instead of one aggregated object (same
          as Accept: application/x-ndjson)'
        in: query
        name: stream
        required: false
        schema:
          enum:
          - "1"
          - "true"
          - "yes"
          type: string
      - description: application/x-ndjson streams per-item results as they complete
        in: header
        name: Accept
        required: false
        schema:
          type: string
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyBatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyBatchResponse'
            application/x-ndjson:
              schema:
                description: 'one JSON object per line: each result tagged with its
                  index, then a final {total, succeeded, failed} summary line'
                type: string
          description: Per-item results in request order, plus a summary
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: Verify many signatures
      tags:
      - batch
...
  /verify/registry:
    post:
      operationId: verifyRegistry
      parameters:
      - description: preferred language for error messages (ru, kk; anything else
          renders English)
        in: header
        name: Accept-Language
        required: false
        schema:
          type: string
      - description: replay the first successful response for this key instead of
          re-executing, so a retry does not produce a second signature
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRegistryRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryResponse'
          description: Register over the whole set
          headers:
            Idempotency-Replayed:
              description: present as true when this response was replayed from an
                earlier identical call
              schema:
                enum:
                - "true"
                type: string
        "400":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        "503":
          $ref: '#/components/responses/Error'
      summary: 'Verify a set of documents and return the auditor''s register: counts
        plus one row per document with its verdict, signers and first failing check'
      tags:
      - batch
servers:
- description: Default REST work port (-http :8080)
  url: http://localhost:8080
...